Revert "✨ feat(data): add `Observation` batch design and refactor format conversion"
