+++ library/docs/design/data/overview.md
@@ -5,8 +5,8 @@ graph TD
     A["action_trainer"]
     A --> B["data/ <br/> Data"]
     B --> C["__init__.py"]
-    B --> D["dataset.py"]
-    B --> E["observation.py"]
+    B --> D["action.py"]
+    B --> E["dataclasses.py"]
     B --> F["datamodules.py"]
     B --> G["gym.py"]
     B --> H["lerobot.py"]
