+++ library/src/getiaction/data/dataclasses.py
@@ -1,0 +1,82 @@
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+"""Types and internal representations."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from enum import StrEnum
+from typing import TYPE_CHECKING, Any
+
+if TYPE_CHECKING:
+    import numpy as np
+    import torch
+
+
+@dataclass(frozen=True)
+class Observation:
+    """A single observation from an imitation learning dataset."""
+
+    # Core Observation
+    action: dict[str, torch.Tensor | np.ndarray] | torch.Tensor | np.ndarray | None = None
+    task: dict[str, torch.Tensor | np.ndarray] | torch.Tensor | np.ndarray | None = None
+    state: dict[str, torch.Tensor | np.ndarray] | torch.Tensor | np.ndarray | None = None
+    images: dict[str, torch.Tensor | np.ndarray] | torch.Tensor | np.ndarray | None = None
+
+    # Optional RL & Metadata Fields
+    next_reward: torch.Tensor | np.ndarray | None = None
+    next_success: bool | None = None
+    episode_index: torch.Tensor | np.ndarray | None = None
+    frame_index: torch.Tensor | np.ndarray | None = None
+    index: torch.Tensor | np.ndarray | None = None
+    task_index: torch.Tensor | np.ndarray | None = None
+    timestamp: torch.Tensor | np.ndarray | None = None
+    info: dict[str, Any] | None = None
+    extra: dict[str, Any] | None = None
+
+    class ComponentKeys(StrEnum):
+        """Enum for batch observation components."""
+
+        STATE = "state"
+        ACTION = "action"
+        IMAGES = "images"
+
+        NEXT_REWARD = "next_reward"
+        NEXT_SUCCESS = "next_success"
+        EPISODE_INDEX = "episode_index"
+        FRAME_INDEX = "frame_index"
+        INDEX = "index"
+        TASK_INDEX = "task_index"
+        TIMESTAMP = "timestamp"
+        INFO = "info"
+        EXTRA = "extra"
+
+
+class FeatureType(StrEnum):
+    """Enum for feature types."""
+
+    VISUAL = "VISUAL"
+    ACTION = "ACTION"
+    STATE = "STATE"
+    ENV = "ENV"
+
+
+@dataclass(frozen=True)
+class Feature:
+    """A feature representation."""
+
+    normalization_data: NormalizationParameters | None = None
+    ftype: FeatureType | None = None
+    shape: tuple[int, ...] | None = None
+    name: str | None = None
+
+
+@dataclass(frozen=True)
+class NormalizationParameters:
+    """Parameters for normalizing a tensor."""
+
+    mean: torch.Tensor | np.ndarray | None = None
+    std: torch.Tensor | np.ndarray | None = None
+    min: torch.Tensor | np.ndarray | None = None
+    max: torch.Tensor | np.ndarray | None = None
