+++ library/src/getiaction/data/lerobot/converters.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 from enum import StrEnum
 from typing import TYPE_CHECKING, Any, cast
 
-from getiaction.data.observation import Observation
+from getiaction.data import Observation
 
 if TYPE_CHECKING:
     import torch
