+++ library/src/getiaction/data/lerobot/dataset.py
@@ -9,16 +9,13 @@ from typing import TYPE_CHECKING, Any
 
 from lightning_utilities import module_available
 
-from getiaction.data.dataset import Dataset
-from getiaction.data.observation import Feature, FeatureType, NormalizationParameters, Observation
+from getiaction.data import Dataset, Feature, FeatureType, NormalizationParameters, Observation
 
 from .converters import FormatConverter
 
 if TYPE_CHECKING:
     from collections.abc import Callable
     from pathlib import Path
-
-    from getiaction.data import Observation
 
 if TYPE_CHECKING or module_available("lerobot"):
     from lerobot.datasets.lerobot_dataset import LeRobotDataset
