+++ library/src/getiaction/policies/lerobot/universal.py
@@ -27,8 +27,6 @@ if TYPE_CHECKING:
     from lerobot.configs.types import PolicyFeature
     from lerobot.policies.pretrained import PreTrainedPolicy
 
-    from getiaction.data import Observation
-
 if TYPE_CHECKING or module_available("lerobot"):
     from lerobot.datasets.utils import dataset_to_policy_features
     from lerobot.policies.factory import get_policy_class, make_policy_config
@@ -341,7 +339,7 @@ class LeRobotPolicy(Policy, LeRobotFromConfig):
         # Initialize policy now
         self._initialize_policy(features, features, self._provided_config, stats)
 
-    def forward(self, batch: Observation) -> torch.Tensor:
+    def forward(self, batch: dict[str, torch.Tensor]) -> torch.Tensor:
         """Forward pass through the LeRobot policy.
 
         Args:
@@ -351,11 +349,11 @@ class LeRobotPolicy(Policy, LeRobotFromConfig):
             Policy output (format depends on policy type).
         """
         # Convert to LeRobot format if needed (handles Observation or collated dict)
-        batch_dict = FormatConverter.to_lerobot_dict(batch)
+        batch = FormatConverter.to_lerobot_dict(batch)
 
-        return self.lerobot_policy.forward(batch_dict)
+        return self.lerobot_policy.forward(batch)
 
-    def training_step(self, batch: Observation, batch_idx: int) -> torch.Tensor:
+    def training_step(self, batch: dict[str, torch.Tensor], batch_idx: int) -> torch.Tensor:
         """Training step for Lightning.
 
         Args:
@@ -368,9 +366,9 @@ class LeRobotPolicy(Policy, LeRobotFromConfig):
         del batch_idx  # Unused argument
 
         # Convert to LeRobot format if needed (handles Observation or collated dict)
-        batch_dict = FormatConverter.to_lerobot_dict(batch)
+        batch = FormatConverter.to_lerobot_dict(batch)
 
-        output = self.lerobot_policy.forward(batch_dict)
+        output = self.lerobot_policy.forward(batch)
 
         # Handle different output formats from LeRobot policies
         if isinstance(output, tuple) and len(output) == 2:  # noqa: PLR2004
@@ -392,7 +390,7 @@ class LeRobotPolicy(Policy, LeRobotFromConfig):
         self.log("train/loss", loss, prog_bar=True)
         return loss
 
-    def validation_step(self, batch: Observation, batch_idx: int) -> torch.Tensor:
+    def validation_step(self, batch: dict[str, torch.Tensor], batch_idx: int) -> torch.Tensor:
         """Validation step for Lightning.
 
         Args:
@@ -405,9 +403,9 @@ class LeRobotPolicy(Policy, LeRobotFromConfig):
         del batch_idx  # Unused argument
 
         # Convert to LeRobot format if needed (handles Observation or collated dict)
-        batch_dict = FormatConverter.to_lerobot_dict(batch)
+        batch = FormatConverter.to_lerobot_dict(batch)
 
-        output = self.lerobot_policy.forward(batch_dict)
+        output = self.lerobot_policy.forward(batch)
 
         # Handle different output formats
         if isinstance(output, tuple) and len(output) == 2:  # noqa: PLR2004
@@ -428,7 +426,7 @@ class LeRobotPolicy(Policy, LeRobotFromConfig):
         self.log("val/loss", loss, prog_bar=True)
         return loss
 
-    def select_action(self, batch: Observation) -> torch.Tensor:
+    def select_action(self, batch: dict[str, torch.Tensor]) -> torch.Tensor:
         """Select action (inference mode) through LeRobot.
 
         Args:
@@ -437,8 +435,7 @@ class LeRobotPolicy(Policy, LeRobotFromConfig):
         Returns:
             Predicted actions.
         """
-        batch_dict = FormatConverter.to_lerobot_dict(batch)
-        return self.lerobot_policy.select_action(batch_dict)
+        return self.lerobot_policy.select_action(batch)
 
     def configure_optimizers(self) -> torch.optim.Optimizer:
         """Configure optimizer for Lightning.
